requirements.txt
click
fastapi
h11
psycopg2-binary
pydantic
python-dotenv
SQLAlchemy
starlette
typing-extensions
uvicorn
pyjwt
passlib
redis
python-multipart